refactor: change BottomSheetFooter package and update imports in related components
