+++ ui/src/main/java/com/baghdad/ui/feature/component/bottomSheet/SavedListBottomSheet.kt
@@ -20,7 +20,6 @@ import com.baghdad.design_system.shared.Selectable
 import com.baghdad.design_system.theme.NovixTheme
 import com.baghdad.ui.R
 import com.baghdad.ui.feature.search.component.BottomSheetHeader
-import com.baghdad.ui.feature.search.component.filter.BottomSheetFooter
 import com.baghdad.viewmodel.share.ListUiState
 
 @Composable
